backend/requirements-dev.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
django-environ